requirements.txt.txt
tensorflow
pillow
numpy
keras
